feat(auth): JWT refresh + OAuth 2.0 (#33)
Ugrade the HTTP error filter, so now we have nice errors, with the micro
service of origin.

Update the login and signup DTO and we only return JWT + Refresh Token,
user infos are in the JWT Token.

ADD the refresh route for refresh the JWT Token with the refresh token.

For the refresh token, the Gateway is validating the refresh token and
send ID + Email to Auth-service for new JWT Tokens.

---------

Co-authored-by: Neven Le Gall <neven.legall@reseau.sncf.fr>
Co-authored-by: Antony David <antonydavid945@gmail.com>