fix(gateway): add pagination in endpoint (#156)
This pull request introduces enhancements to the `ProjectsController` in
the `apps/api-gateway` project, focusing on authentication, API
documentation, and pagination for the `findByJWTToken` method. The most
important changes include adding guards for JWT-based authentication,
improving API documentation with additional metadata, and implementing
pagination support.

### Authentication Enhancements:
* Added `AuthGuard` using the `jwt` strategy to secure the
`findByJWTToken` method.
* Imported `AuthGuard` from `@nestjs/passport` and added
`@ApiBearerAuth` to indicate that the endpoint requires a bearer token
for authentication.
[[1]](diffhunk://#diff-76e749b4b8bdfa21a9059ba0cc87daf4d675caf2e403035a6439f7867b5303eaR13-R19)
[[2]](diffhunk://#diff-76e749b4b8bdfa21a9059ba0cc87daf4d675caf2e403035a6439f7867b5303eaL127-R142)

### API Documentation Improvements:
* Updated the `@ApiOperation` decorator to provide a more detailed
summary of the `findByJWTToken` method, including information about
paginated results for teachers and students.
* Added `@ApiQuery` decorators for `page` and `size` parameters to
document optional pagination inputs.

### Pagination Support:
* Modified the `findByJWTToken` method to accept `page` and `size` query
parameters, with default values of `1` and `10`, respectively, and
forwarded these parameters to the proxy request.